DDCNLS-374: Fixing tests broken due to refactoring of DES model classes.
